refactor: rename FloatingService as `ControllerServce`
- Some code styles are refined as well.

Fixes https://github.com/thumbs-project/thumbs-project/issues/20
