+++ app/src/main/java/com/thumbs/android/thumbsAndroid/views/MainActivity.kt
@@ -8,19 +8,17 @@ import android.provider.Settings
 import android.support.v7.app.AppCompatActivity
 import android.widget.Toast
 import com.thumbs.android.thumbsAndroid.R
-import com.thumbs.android.thumbsAndroid.services.FloatingViewService
+import com.thumbs.android.thumbsAndroid.services.ControllerService
 import kotlinx.android.synthetic.main.activity_main.*
 
 
 class MainActivity : AppCompatActivity() {
-
   val PERMISSION_CODE = 2002
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.activity_main)
     init()
-
     buttonSetting.setOnClickListener { startActivity(Intent (settingActivityIntent())) }
   }
 
@@ -28,11 +26,11 @@ class MainActivity : AppCompatActivity() {
     buttonCreateWidget.setOnClickListener {
       when {
         Build.VERSION.SDK_INT < Build.VERSION_CODES.M -> {
-          startService(Intent(this@MainActivity, FloatingViewService::class.java))
+          startService(Intent(this@MainActivity, ControllerService::class.java))
           finish()
         }
         Settings.canDrawOverlays(this@MainActivity) -> {
-          startService(Intent(this@MainActivity, FloatingViewService::class.java))
+          startService(Intent(this@MainActivity, ControllerService::class.java))
           finish()
         }
         else -> {
@@ -40,10 +38,7 @@ class MainActivity : AppCompatActivity() {
           Toast.makeText(this, "You need System Alert Window Permission to do this", Toast.LENGTH_SHORT).show()
         }
       }
-
-
     }
-
   }
 
   private fun checkPermission() {
